requirements.txt
acme==1.17.0
certbot==1.17.0
certbot-nginx==1.17.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
ConfigArgParse==1.5.1
configobj==5.0.6
cryptography==3.4.7
distro==1.5.0
# Editable install with no version control (dns-udns==0.0.0)
-e /home/rob/udns/plugins
idna==3.2
josepy==1.8.0
parsedatetime==2.6
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
ultra-rest-client==0.1.4
urllib3==1.26.6
zope.component==5.0.1
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
